watch_state and master_state exit handling
Beforehead, once exit, these two states are meant to transit to the
master_closing_state and watch_closed_state respectively in regard to
the zk connection close.

However, closing zk connection will immediately cause the session to
expire. That will cause that the watcher promote to master right after
old master close connection, but before old master finished gracefully
stopped.

So we delayed the closing of client connection as the last step.
Accordingly, we allow "watch_state -> exit_state" transition.

As for master_state, it is a little tricky. Once it received exit event
in master_state, it will firstly transition to master_closing_state.
If it finishes closing exceeds timeout, then it just transit to
exit_state; Otherwise, it will transit to master_closed_state, but in
this case, it will directly get a second "exit" event from `exitCh` and
directly to exit_state. So the tricky here is that once we get a exit
signal, we will send 2 "exit" events to `exitCh`.